Moved Slot Component Interface into Journal Components
