+++ src/pages/journal/journal-components/slot/slot.ts
